+++ routes/api/users/users.controller.js
@@ -12,6 +12,7 @@
 const Users = require('./users.scheme')
 const passwordHash = require('password-hash')
 const createError = require('http-errors')
+const { validationResult } = require('express-validator')
 
 exports.login = (username, password) => {
     return new Promise((resolve, reject) => {
@@ -31,7 +32,6 @@ exports.login = (username, password) => {
     })
 }
 
-
 exports.findAll = (req, res, next) => {
     const q = req.query;
     const where = {}
@@ -40,12 +40,17 @@ exports.findAll = (req, res, next) => {
     if (q.username) where['username'] = q.username
     if (q.displayName) where['displayName'] = q.displayName
     
-    Users.find(where).limit(req.query.limit || 0).skip(req.query.skip || 0).populate('role').then(users => {
+    Users.find(where).limit(req.query.limit || 0).skip(req.query.skip || 0).populate('role').sort('-createdAt').then(users => {
         res.json(users)
     }).catch(e => next(e))
 }
 
 exports.findById = (req, res, next) => {
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+        return res.status(422).json({ errors: errors.array() });
+    }
+
     const id = req.params.id
 
     Users.findById(id).populate('role').then(users => {
@@ -54,6 +59,11 @@ exports.findById = (req, res, next) => {
 }
 
 exports.insert = (req, res, next) => {
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+        return res.status(422).json({ errors: errors.array() })
+    }
+
     const data = req.body;
 
     data.password = passwordHash.generate(data.password)
@@ -66,6 +76,11 @@ exports.insert = (req, res, next) => {
 }
 
 exports.update = (req, res, next) => {
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+        return res.status(422).json({ errors: errors.array() })
+    }
+    
     const id = req.params.id
     const data = req.body;
 
@@ -80,6 +95,11 @@ exports.update = (req, res, next) => {
 }
 
 exports.removeById = (req, res, next) => {
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+        return res.status(422).json({ errors: errors.array() });
+    }
+
     const id = req.params.id
     
     Users.findByIdAndRemove(id).then(users => {
@@ -97,4 +117,9 @@ exports.remove = (req, res, next) => {
             data: users
         })
     }).catch(e => next(e))
+}
+
+//checking if user or email has been registered
+exports.findByUserOrEmail = (value) => {
+    return Users.findOne({ $or: [{ username: value }, { email: value }] })
 }
